update: remove nested links folder
So, the examples folder has two primary uses today: validation of the
SDKs and payload examples for users.

Due to the first case of validation it makes adding examples for links
very difficult in that you can't validate against the links without
having other events. So either those would need to be mocked, etc. We'll
need to figure that out at some point but that's out of scope for the
PR. So this commit instead removes the nested links folder in examples,
and instead just adds the `chain_id` in the context of all CDEvents.

Signed-off-by: benjamin-j-powell <bjp@apple.com>